# docs/swagger.yaml
definitions:
...
  apiSponsor.RecordAddRequest:
    properties:
      amount:
        description: 赞助金额(分)
        example: 1000
        minimum: 1
        type: integer
      channel_id:
        description: 赞助渠道 ID
        example: 123456789
        type: integer
      is_anonymous:
        description: 是否匿名
        example: false
        type: boolean
      is_hidden:
        description: 是否隐藏
        example: false
        type: boolean
      message:
        description: 留言信息
        example: 感谢开发者
        maxLength: 500
        type: string
      nickname:
        description: 赞助者昵称
        example: 张三
        maxLength: 100
        minLength: 1
        type: string
      redirect_url:
        description: 跳转链接
        example: https://example.com
        maxLength: 500
        type: string
      sort_order:
        description: 排序值
        example: 0
        minimum: 0
        type: integer
      sponsor_at:
        description: 赞助时间
        example: "2025-01-01T12:00:00Z"
        type: string
    required:
    - amount
    - nickname
    type: object
  apiSponsor.RecordAddResponse:
    properties:
      amount:
        type: integer
      channel:
        $ref: '#/definitions/apiSponsor.SponsorChannelSimpleResponse'
      channel_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_anonymous:
        type: boolean
      is_hidden:
        type: boolean
      message:
        type: string
      nickname:
        type: string
      redirect_url:
        type: string
      sort_order:
        type: integer
      sponsor_at:
        type: string
      updated_at:
        type: string
    type: object
  apiSponsor.RecordDeleteResponse:
    properties:
      message:
        type: string
    type: object
  apiSponsor.RecordDetailResponse:
    properties:
      amount:
        type: integer
      channel:
        $ref: '#/definitions/apiSponsor.SponsorChannelSimpleResponse'
      channel_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_anonymous:
        type: boolean
      is_hidden:
        type: boolean
      message:
        type: string
      nickname:
        type: string
      redirect_url:
        type: string
      sort_order:
        type: integer
      sponsor_at:
        type: string
      updated_at:
        type: string
    type: object
  apiSponsor.RecordEntityResponse:
    properties:
      amount:
        type: integer
      channel:
        $ref: '#/definitions/apiSponsor.SponsorChannelSimpleResponse'
      channel_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_anonymous:
        type: boolean
      is_hidden:
        type: boolean
      message:
        type: string
      nickname:
        type: string
      redirect_url:
        type: string
      sort_order:
        type: integer
      sponsor_at:
        type: string
      updated_at:
        type: string
    type: object
  apiSponsor.RecordPageResponse:
    properties:
      data:
        description: 数据列表
        items:
          $ref: '#/definitions/apiSponsor.RecordEntityResponse'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/base.PaginationInfo'
        description: 分页信息
    type: object
  apiSponsor.RecordPublicItemResponse:
    properties:
      amount:
        type: integer
      channel:
        $ref: '#/definitions/apiSponsor.SponsorChannelSimpleResponse'
      id:
        type: integer
      message:
        type: string
      nickname:
        type: string
      redirect_url:
        type: string
      sponsor_at:
        type: string
    type: object
  apiSponsor.RecordPublicPageResponse:
    properties:
      data:
        description: 数据列表
        items:
          $ref: '#/definitions/apiSponsor.RecordPublicItemResponse'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/base.PaginationInfo'
        description: 分页信息
    type: object
  apiSponsor.RecordUpdateRequest:
    properties:
      amount:
        description: 赞助金额(分)
        example: 1000
        minimum: 1
        type: integer
      channel_id:
        description: 赞助渠道 ID
        example: 123456789
        type: integer
      is_anonymous:
        description: 是否匿名
        example: false
        type: boolean
      is_hidden:
        description: 是否隐藏
        example: false
        type: boolean
      message:
        description: 留言信息
        example: ""
        maxLength: 500
        type: string
      nickname:
        description: 赞助者昵称
        example: 张三
        maxLength: 100
        minLength: 1
        type: string
      redirect_url:
        description: 跳转链接
        example: https://example.com
        maxLength: 500
        type: string
      sort_order:
        description: 排序值
        example: 0
        minimum: 0
        type: integer
      sponsor_at:
        description: 赞助时间
        example: ""
        type: string
    type: object
  apiSponsor.RecordUpdateResponse:
    properties:
      amount:
        type: integer
      channel:
        $ref: '#/definitions/apiSponsor.SponsorChannelSimpleResponse'
      channel_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_anonymous:
        type: boolean
      is_hidden:
        type: boolean
      message:
        type: string
      nickname:
        type: string
      redirect_url:
        type: string
      sort_order:
        type: integer
      sponsor_at:
        type: string
      updated_at:
        type: string
    type: object
  apiSponsor.SponsorChannelSimpleResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  base.PaginationInfo:
    properties:
      has_next:
        description: 是否有下一页
        type: boolean
      has_prev:
        description: 是否有上一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/sponsors/records:
    get:
      consumes:
      - application/json
      description: 分页获取赞助记录列表，支持过滤和排序
      parameters:
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10，最大100）
        in: query
        name: page_size
        type: integer
      - description: 渠道ID过滤
        format: int64
        in: query
        name: channel_id
        type: integer
      - description: 昵称模糊搜索
        in: query
        name: nickname
        type: string
      - description: 是否匿名过滤
        in: query
        name: is_anonymous
        type: boolean
      - description: 是否隐藏过滤
        in: query
        name: is_hidden
        type: boolean
      - description: 排序字段（nickname, amount, sponsor_at, sort_order, created_at）
        in: query
        name: order_by
        type: string
      - description: 排序方向（asc, desc）
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiSponsor.RecordPageResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取赞助记录分页列表
      tags:
      - 赞助记录管理
    post:
      consumes:
      - application/json
      description: 创建新的赞助记录
      parameters:
      - description: 添加赞助记录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiSponsor.RecordAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/apiSponsor.RecordAddResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 添加赞助记录
      tags:
      - 赞助记录管理
  /api/v1/admin/sponsors/records/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的赞助记录
      parameters:
      - description: 赞助记录ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/apiSponsor.RecordDeleteResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 赞助记录不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 删除赞助记录
      tags:
      - 赞助记录管理
    get:
      consumes:
      - application/json
      description: 根据ID获取指定赞助记录的详细信息
      parameters:
      - description: 赞助记录ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiSponsor.RecordDetailResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 赞助记录不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取赞助记录详情
      tags:
      - 赞助记录管理
    put:
      consumes:
      - application/json
      description: 更新指定赞助记录的信息
      parameters:
      - description: 赞助记录ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新赞助记录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiSponsor.RecordUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/apiSponsor.RecordUpdateResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 赞助记录不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新赞助记录
      tags:
      - 赞助记录管理
...
  /api/v1/sponsors/records:
    get:
      consumes:
      - application/json
      description: 分页获取前台赞助墙展示的记录列表，只返回未隐藏的记录，匿名记录显示为"匿名用户"
      parameters:
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认20，最大50）
        in: query
        name: page_size
        type: integer
      - description: 渠道ID过滤
        format: int64
        in: query
        name: channel_id
        type: integer
      - description: 排序字段（amount, sponsor_at, sort_order）
        in: query
        name: order_by
        type: string
      - description: 排序方向（asc, desc）
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiSponsor.RecordPublicPageResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取赞助记录公开分页列表
      tags:
      - 赞助记录公开接口
swagger: "2.0"